Add Frame and Fix BIG problem
The name in Title was changed to Diasoria, and a BIG problem was also solved with the fact that the folder where the pages were located was named Page, which caused a conflict that broke the project
For the future, before committing and branch, launch the project, even if there were very few changes
