+++ MainWindow.xaml.cs
@@ -1,17 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using Diasoria.Pages;
 using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Navigation;
-using System.Windows.Shapes;
 
 namespace Diasoria
 {
@@ -23,6 +11,7 @@ namespace Diasoria
         public MainWindow()
         {
             InitializeComponent();
+            MainFrame.Navigate(new CreationLoadProjectPage());
         }
     }
 }
