+++ Pages/CreationLoadProjectPage.xaml
@@ -1,12 +1,12 @@
-﻿<Page x:Class="Diasoria.Page.ProjectEditorPage"
+﻿<Page x:Class="Diasoria.Pages.CreationLoadProjectPage"
       xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
       xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
       xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
       xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-      xmlns:local="clr-namespace:Diasoria.Page"
+      xmlns:local="clr-namespace:Diasoria.Pages"
       mc:Ignorable="d" 
       d:DesignHeight="450" d:DesignWidth="800"
-      Title="ProjectEditorPage">
+      Title="CreationLoadProjectPage" Background="Blue">
 
     <Grid>
