+++ Pages/CreationLoadProjectPage.xaml.cs
@@ -1,0 +1,15 @@
+﻿using System.Windows.Controls;
+
+namespace Diasoria.Pages
+{
+    /// <summary>
+    /// Логика взаимодействия для CreationLoadProjectPage.xaml
+    /// </summary>
+    public partial class CreationLoadProjectPage : Page
+    {
+        public CreationLoadProjectPage()
+        {
+            InitializeComponent();
+        }
+    }
+}
