+++ Pages/ProjectEditorPage.xaml.cs
@@ -1,0 +1,15 @@
+﻿using System.Windows.Controls;
+
+namespace Diasoria.Pages
+{
+    /// <summary>
+    /// Логика взаимодействия для ProjectEditorPage.xaml
+    /// </summary>
+    public partial class ProjectEditorPage : Page
+    {
+        public ProjectEditorPage()
+        {
+            InitializeComponent();
+        }
+    }
+}
